+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `SQLiteDatabase#delete()` method
 - `SQLiteDatabase#insert()` method
+- `SQLiteDatabase#replace()` method
 - `SQLiteDatabase#update()` method
 
 ## [0.2.2] - 2018-12-04